Remove local json database for workouts
