+++ src/database/Workout.js
@@ -1,63 +1,35 @@
+const mongoose = require('mongoose');
+mongoose.connect('mongodb://localhost:27017/crossfit');
 
-/**
- * @openapi
- * components:
- *   schemas:
- *     Workout:
- *       type: object
- *       properties:
- *         id: 
- *           type: string
- *           example: 61dbae02-c147-4e28-863c-db7bd402b2d6
- *         name: 
- *           type: string
- *           example: Tommy V  
- *         mode:
- *           type: string
- *           example: For Time
- *         equipment:
- *           type: array
- *           items:
- *             type: string
- *           example: ["barbell", "rope"]
- *         exercises:
- *           type: array
- *           items:
- *             type: string
- *           example: ["21 thrusters", "12 rope climbs, 15 ft", "15 thrusters", "9 rope climbs, 15 ft", "9 thrusters", "6 rope climbs, 15 ft"]
- *         createdAt:
- *           type: string
- *           example: 4/20/2022, 2:21:56 PM
- *         updatedAt: 
- *           type: string
- *           example: 4/20/2022, 2:21:56 PM
- *         trainerTips:
- *           type: array
- *           items:
- *             type: string
- *           example: ["Split the 21 thrusters as needed", "Try to do the 9 and 6 thrusters unbroken", "RX Weights: 115lb/75lb"]
- */
-const DB = require("./db.json");
-const { saveToDatabase } = require("./utils");
+const Workouts = mongoose.model('Workouts', {
+    name: String,
+    mode: String,
+    equipment: [String],
+    exercises: [String],
+    createdAt: { type: Date, default: Date.now },
+    updatedAt: { type: Date, default: Date.now },
+    trainerTips: [String]
+});
 
-const getAllWorkouts = (filterParams) => {
+const createNewWorkout = (newWorkout) => {
     try {
-        let workouts = DB.workouts;
-        if (filterParams.mode) {
-            return DB.workouts.filter((workout) =>
-                workout.mode.toLowerCase().includes(filterParams.mode)
-            );
-        } 
-        // Other if-statements will go here for different parameters
-        return workouts;
+        const workout = new Workouts({
+            name: newWorkout.name,
+            mode: newWorkout.mode,
+            equipment: newWorkout.equipment,
+            exercises: newWorkout.exercises,
+            trainerTips: newWorkout.trainerTips
+        });
+        workout.save();
     } catch (error) {
-        throw { status: 500, message: error };
+        throw { status: error?.status || 500, message: error?.message || error };
     }
 };
 
-const getOneWorkout = (workoutId) => {
+
+const getOneWorkout = async (workoutId) => {
     try {
-        const workout = DB.workouts.find((workout) => workout.id === workoutId);
+        const workout = await Workouts.findById(workoutId, { __v: 0 })
         if (!workout) {
             throw {
                 status: 404,
@@ -70,79 +42,70 @@ const getOneWorkout = (workoutId) => {
     }
 };
 
-const createNewWorkout = (newWorkout) => {
+
+const getAllWorkouts = async (filterParams) => {
     try {
-        const isAlreadyAdded =
-            DB.workouts.findIndex((workout) => workout.name === newWorkout.name) > -1;
-        if (isAlreadyAdded) {
-            throw {
-                status: 400,
-                message: `Workout with the name '${newWorkout.name}' already exists`,
-            };
-        }
-        DB.workouts.push(newWorkout);
-        saveToDatabase(DB);
-        return newWorkout;
+        return await Workouts.find({ __v: 0 });
     } catch (error) {
-        throw { status: error?.status || 500, message: error?.message || error };
+        throw { status: 500, message: error };
     }
 };
 
-const updateOneWorkout = (workoutId, changes) => {
+
+const updateOneWorkout = async (workoutId, changes) => {
     try {
-        const isAlreadyAdded =
-            DB.workouts.findIndex((workout) => workout.name === changes.name) > -1;
-        if (isAlreadyAdded) {
+        const workoutByName = await Workouts.findOne({ 'name': changes.name });
+        if (workoutByName) {
             throw {
                 status: 400,
                 message: `Workout with the name '${changes.name}' already exists`,
             };
         }
-        const indexForUpdate = DB.workouts.findIndex(
-            (workout) => workout.id === workoutId
-        );
-        if (indexForUpdate === -1) {
+        const workoutById = await Workouts.findById(workoutId)
+        if (!workoutById) {
             throw {
-                status: 400,
+                status: 404,
                 message: `Can't find workout with the id '${workoutId}'`,
             };
         }
-        const updatedWorkout = {
-            ...DB.workouts[indexForUpdate],
-            ...changes,
+
+        var res = await Workouts.findByIdAndUpdate({ _id: workoutId }, {
+            name: changes.name,
+            mode: changes.mode,
+            equipment: changes.equipment,
+            exercises: changes.exercises,
+            trainerTips: changes.trainerTips,
             updatedAt: new Date().toLocaleString("en-US", { timeZone: "UTC" }),
-        };
-        DB.workouts[indexForUpdate] = updatedWorkout;
-        saveToDatabase(DB);
-        return updatedWorkout;
+        })
+
+        return res;
     } catch (error) {
         throw { status: error?.status || 500, message: error?.message || error };
     }
 };
 
-const deleteOneWorkout = (workoutId) => {
+
+const deleteOneWorkout = async (workoutId) => {
     try {
-        const indexForDeletion = DB.workouts.findIndex(
-            (workout) => workout.id === workoutId
-        );
-        if (indexForDeletion === -1) {
+
+        const workoutById = await Workouts.findById(workoutId)
+        if (!workoutById) {
             throw {
-                status: 400,
+                status: 404,
                 message: `Can't find workout with the id '${workoutId}'`,
             };
         }
-        DB.workouts.splice(indexForDeletion, 1);
-        saveToDatabase(DB);
+        await Workouts.findByIdAndDelete(workoutId)
+
     } catch (error) {
         throw { status: error?.status || 500, message: error?.message || error };
     }
 };
-
 
 module.exports = {
-    getAllWorkouts,
     createNewWorkout,
     getOneWorkout,
+    getAllWorkouts,
     updateOneWorkout,
-    deleteOneWorkout,
+    deleteOneWorkout
 };
